Задание на 5+ перенесено в отдельный класс
