+++ src/test/java/com/sbrf/reboot/atm/cassettes/CassetteTest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
 
 class CassetteTest {
 
@@ -28,29 +26,5 @@ class CassetteTest {
         }});
 
         Assertions.assertEquals(1, cassette.getCountBanknotes());
-    }
-}
-
-/**
- * Задание на 5+.
- */
-class AdvancedTest {
-    abstract class Vehicle {
-    }
-
-    abstract class Car extends Vehicle {
-    }
-
-    public class ObjectStore<K, V> {
-        void putAll(Map<? extends K, ? extends V> entries) {
-
-        }
-    }
-
-    public void main(String[] args) {
-        ObjectStore<String, Vehicle> store = new ObjectStore<>();
-
-        Map<String, Car> cars = new HashMap<>();
-        store.putAll(cars);
     }
 }
